refactor: create grouped messages
