+++ src/Components/ChatComponent/components/GroupedMessages/GroupedMessages.css
@@ -1,0 +1,74 @@
+.message {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 10px;
+}
+
+.message.received {
+  align-items: flex-start;
+  padding-right: 100px;
+}
+
+.message .text {
+  padding: 10px;
+  border-radius: 15px;
+  background-color: #e0e0e0;
+  color: #333;
+}
+
+.message .text img {
+  max-height: 65vh;
+}
+
+.message-content {
+  display: flex;
+  flex-direction: column;
+}
+
+.message-row {
+  display: flex;
+  align-items: center;
+}
+
+.text img {
+  cursor: pointer;
+  width: auto;
+  background-color: white;
+  color: black;
+}
+
+.message-time {
+  font-size: 0.75em;
+  color: white;
+  margin-top: 3px;
+  text-align: right;
+}
+
+.client-name {
+  color: #0f824c77;
+  font-weight: bold;
+}
+
+.reaction-toggle-button {
+  display: flex;
+  cursor: pointer;
+  align-items: center;
+  justify-content: flex-end;
+  font-size: 18px;
+}
+
+.time-messages {
+  color: #333;
+  font-weight: bold;
+}
+
+.reaction-buttons {
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
+
+.reaction-buttons div {
+  cursor: pointer;
+  font-size: 18px;
+}
