+++ src/Components/ChatComponent/components/GroupedMessages/index.js
@@ -1,0 +1,1 @@
+export * from "./GroupedMessages"
