+++ src/Components/ChatComponent/components/index.js
@@ -2,3 +2,4 @@ export * from "./ChatListItem"
 export * from "./Media"
 export * from "./Merge"
 export * from "./PersonalData4ClientForm"
+export * from "./GroupedMessages"
